Added several instances of error checking. Updated front end visuals. Fixed a few broken features. Added visualization of users name when logged in. Created links from batches to their respective recipes.

+++ resources/views/recipes/showIndividual.blade.php
@@ -10,33 +10,42 @@
 
 @section('content')
 
-    @foreach($recipes as $recipe)
-        <h1>{{$recipe['recipe_name']}}</h1>
-        Recipe Name: {{ $recipe['recipe_name'] }}
+    @if(empty($recipe))
+        That recipe does not exist.
         <br>
-        Honey Type: {{ $recipe['honey_type'] }}
+        It may have been deleted, or never existed at all.
         <br>
-        Yeast Type: {{ $recipe['yeast_type'] }}
-        <br>
-        Difficulty: {{ $recipe['difficulty'] }}
-        <br>
-        <br>
-        <div class='form-group'>
-            <label for='recipe_text'>* Recipe Text:</label>
-                <textarea
+        Why not <a href="/recipes/create"><strong>create one</strong>?</a>
+
+    @else
+            <h1>{{$recipe['recipe_name']}}</h1>
+            Recipe Name: {{ $recipe['recipe_name'] }}
+            <br>
+            Honey Type: {{ $recipe['honey_type'] }}
+            <br>
+            Yeast Type: {{ $recipe['yeast_type'] }}
+            <br>
+            Difficulty: {{ $recipe['difficulty'] }}
+            <br>
+            <br>
+            <div class='form-group'>
+                <label for='recipe_text'>* Recipe Text:</label>
+                    <textarea
                         style="height: 250px; width: 500px;"
                         name="recipe_text"
                         class="form-control"
                         id="recipe_text" disabled>
-                    {{$recipe->recipe_text}}
-                </textarea>
-        </div>
-        <br>
-    @endforeach
+                        {{$recipe->recipe_text}}
+                    </textarea>
+            </div>
+            <br>
 
-    <h4><a href="/recipes/edit/{{$recipe->id}}">Edit Recipe</a></h4>
+        @if(Auth::id() == $recipe->user_id)
+            <h4><a href="/recipes/edit/{{$recipe->id}}">Edit Recipe</a></h4>
+            <h4><a href="/recipes/confirm-delete/{{$recipe->id}}">Delete Recipe</a></h4>
+        @endif
 
-    <h4><a href="/recipes/confirm-delete/{{$recipe->id}}">Delete Recipe</a></h4>
+    @endif
 
 @stop
